% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_cutpoints.R
\name{boxplot_cutpoints}
\alias{boxplot_cutpoints}
\title{Boxplot Outlier Cut Points}
\usage{
boxplot_cutpoints(x, iqr_mult = 3)
}
\arguments{
\item{x}{a numeric vector.}

\item{iqr_mult}{interquartile range multiplier. Defaults to 3.}
}
\value{
\itemize{
\item{\code{medcouple} }{medcouple value returned from \code{robustbase::mc(x)}.}
\item{\code{standard_rule} }{lower and upper limits.}
\item{\code{adjusted_rule} }{lower and upper limits.}
}
}
\description{
Constructs boxplot style rules for flagging outliers.
}
\details{
Computes two outlier rules. The standard boxplot outlier rule based on \code{Q1 - IQR*iqr_mult} and \code{Q3 + IQR*iqr_mult}. The adjusted boxplot rule which uses the medcouple to account for data skewness. The medcouple depends on the \pkg{robustbase} package.
}
\examples{
x1 = rnorm(1000,0,1)
x2 = rexp(1000,1)
boxplot_cutpoints(x=x1, iqr_mult=1.5)
boxplot_cutpoints(x=x2, iqr_mult=1.5)
}
\references{
M. Hubert; E. Vandervieren (2008). "An adjusted boxplot for skewed distributions". Computational Statistics and Data Analysis. 52 (12): 5186-5201.
}
\seealso{
\code{\link[robustbase]{mc}}, \code{\link{score_threshold}}
}
