% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_output.R
\name{code_output}
\alias{code_output}
\title{SeaClass Output Code Creator}
\usage{
code_output(methodvar, dfvar = NULL, responsevar = NULL, tool1 = NULL,
  tool2 = NULL)
}
\arguments{
\item{methodvar}{current application method variable. Options include: "Bias Correction", "Cost Models", "Decision Tree", "Mahalanobis Distance", "One Class SVM", "Prediction Models", "Single Variable: All", "Single Variable: Categorical", "SMOTE", "Train/Test Split", "Thresholds: Accuracy", "Thresholds: True Positive", "Variable Selection: PCA", and "Variable Selection: Supervised".}

\item{dfvar}{current application data frame variable. Defaults to NULL.}

\item{responsevar}{current application response variable. Defaults to NULL.}

\item{tool1}{current application input variable from shiny \code{input$tool1}. Defaults to NULL.}

\item{tool2}{current application input variable from shiny \code{input$tool2}. Defaults to NULL.}
}
\value{
Returns R code in text format for current application analysis.
}
\description{
SeaClass function which outputs analysis code from the output tab based on users input and actions.
}
\details{
SeaClass helper function; not intended for general use.
}
\examples{
\dontrun{
X <- matrix(rnorm(1000,0,1),ncol=10,nrow=100)
X[1:10,] <- X[1:10,] + 3
Y <- c(rep(1,10), rep(0,990))
DF <- data.frame(Y = Y , X)
my_text <- code_output(methodvar = "Mahalanobis Distance", dfvar = "DF", responsevar = "Y", tool1 = 20, tool2 = NULL)
cat(my_text)}
}
\seealso{
\code{\link{SeaClass}}
}
