% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_check.R
\name{cov_check}
\alias{cov_check}
\title{Covariance Check}
\usage{
cov_check(cov, cut_point = 0.9999)
}
\arguments{
\item{cov}{a matrix.}

\item{cut_point}{threshold which signifies two variables are highly dependent and one should be omitted. Specifically, when a group of variables are highly dependent, only the first variable is kept and all others are dropped. Defaults to 0.9999.}
}
\value{
\itemize{
\item{\code{remove_vars} }{locations of variables to be dropped.}
\item{\code{remove_length} }{number of variables to be dropped.}
\item{\code{remove_text} }{a message detailing the recommended action.}
}
}
\description{
Checks a covariance matrix and identifies variables which should be dropped due to zero variance or high dependence.
}
\details{
Used to flag variables which will likely cause the covariance matrix to be singular or near singular; and hence, non-invertible.
}
\examples{
cov1 <- matrix(c(1,0,0,0,0,0,0,0,0,0,1,1,0,0,1,1),ncol=4,nrow=4)
cov2 <- diag(c(1,1,1,1))
cov_check(cov1, cut_point = 0.9999)
cov_check(cov2, cut_point = 0.9999)
}
\seealso{
\code{\link{cov_diag_check}}, \code{\link{cov_offdiag_check}}
}
