% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_diag_check.R
\name{cov_diag_check}
\alias{cov_diag_check}
\title{Covariance Diagonal Check}
\usage{
cov_diag_check(cov)
}
\arguments{
\item{cov}{a matrix.}
}
\value{
\itemize{
\item{\code{remove_vars} }{locations of zero entries.}
\item{\code{remove_length} }{number of zero entries.}
\item{\code{remove_text} }{a message detailing an action taken by the covariance check wrapper function \code{cov_check()}.}
}
}
\description{
Identifies any zero entries along the diagonal of a matrix.
}
\details{
Intended to be a helper function for flagging zero variance variables whenever the input is a covariance matrix. A covariance matrix containing zeros on the diagonal will be singular; and hence, non-invertible.
}
\examples{
cov1 <- diag(c(1,0,3,0))
cov2 <- diag(1:4)
cov_diag_check(cov1)
cov_diag_check(cov2)
}
\seealso{
\code{\link{cov_check}}
}
