% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_offdiag_check.R
\name{cov_offdiag_check}
\alias{cov_offdiag_check}
\title{Covariance Off Diagonal Check}
\usage{
cov_offdiag_check(cov, cut_point = 0.9999)
}
\arguments{
\item{cov}{a matrix.}

\item{cut_point}{threshold which signifies two variables are highly dependent and one should be omitted. Specifically, when a group of variables are highly dependent, only the first variable is kept and all others are dropped. Defaults to 0.9999.}
}
\value{
\itemize{
\item{\code{remove_vars} }{locations of variables to be dropped.}
\item{\code{remove_length} }{number of variables to be dropped.}
\item{\code{remove_text} }{a message detailing an action taken by the covariance check wrapper function \code{cov_check()}.}
}
}
\description{
Converts a covariance matrix to a correlation matrix and then identifies variables which should be dropped due to high dependence.
}
\details{
Intended to be a helper function for flagging highly dependent variables which cause the covariance matrix to be singular or near singular; and hence, non-invertible.
}
\examples{
cov1 <- matrix(c(1,0,0,0,1,1,0,1,1),ncol=3,nrow=3)
cov2 <- diag(c(1,1,1))
cov_offdiag_check(cov1, cut_point = 0.9999)
cov_offdiag_check(cov2, cut_point = 0.9999)
}
\seealso{
\code{\link{cov_check}}
}
