% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_fp_threshold.R
\name{max_fp_threshold}
\alias{max_fp_threshold}
\title{Best Threshold Rule for a User's Maximum False Positive Rate}
\usage{
max_fp_threshold(x, group, pos_class, max_fp = 0.05)
}
\arguments{
\item{x}{a numeric predictor.}

\item{group}{binary grouping variable.}

\item{pos_class}{group level for positive class.}

\item{max_fp}{maximum false positive rate. Defaults to 0.05.}
}
\value{
\itemize{
\item{\code{FP} }{false positive percentage.}
\item{\code{TP} }{true positive percentage.}
\item{\code{Misclass} }{misclassification percentage.}
\item{\code{Direction} }{text summarizing threshold rule.}
}
}
\description{
Returns the best threshold rule based on true positive rate, where the rule has a false positive rate not exceeding a user specified maximum.
}
\details{
Output quantities for false positive, true positive, and misclassification are percentages.
}
\examples{
x <- c(rnorm(100,0,1),rnorm(100,2,1))
group <- c(rep(0,100),rep(2,100))
max_fp_threshold(x=x, group=group, pos_class=2, max_fp = 0.05)
max_fp_threshold(x=x, group=group, pos_class=0, max_fp = 0.05)
}
\seealso{
\code{\link{accuracy_threshold}}, \code{\link{score_threshold}}
}
