% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rareFailData}
\alias{rareFailData}
\title{Rare Failure Data}
\format{A data frame with 9000 rows and 25 variables:
\describe{
  \item{BinaryResponse}{the test outcome (PASS/FAIL),}
  \item{X1, X2,..., X21}{numeric explanatory variables (standardized),}
  \item{category1, category2, category3}{categorical explanatory variables.}
}}
\usage{
rareFailData
}
\description{
A data set containing a binary variable with few failing cases yet many passing cases. The data set also contains 21 numeric and 3 categorical explanatory variables.
}
\keyword{datasets}
