% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_threshold.R
\name{score_threshold}
\alias{score_threshold}
\title{Score a Binary Threshold}
\usage{
score_threshold(x, group, pos_class, cut, type = "upper")
}
\arguments{
\item{x}{a numeric predictor.}

\item{group}{binary grouping variable.}

\item{pos_class}{group level for positive class.}

\item{cut}{user specified threshold to divide \code{x} input.}

\item{type}{direction of positive group. Options are "upper" (default) and "lower".}
}
\value{
\itemize{
\item{\code{FP} }{false positive percentage.}
\item{\code{TP} }{true positive percentage.}
\item{\code{Misclass} }{misclassification percentage.}
\item{\code{Direction} }{text summarizing threshold rule.}
}
}
\description{
Scores a user specified threshold.
}
\details{
Output quantities for false positive, true positive, and misclassification are percentages.
}
\examples{
x <- c(rnorm(100,0,1),rnorm(100,2,1))
group <- c(rep(0,100),rep(2,100))
score_threshold(x=x, group=group, pos_class=2, cut=1, type="upper")
score_threshold(x=x, group=group, pos_class=0, cut=1, type="lower")
}
\seealso{
\code{\link{accuracy_threshold}}, \code{\link{max_fp_threshold}}
}
