% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeaClass.R
\name{SeaClass}
\alias{SeaClass}
\title{SeaClass Interactive GUI}
\usage{
SeaClass()
}
\value{
None.
}
\description{
The SeaClass function starts up the interactive GUI for binary response variable analysis.
}
\details{
Detailed instructions are included on the application's info tab. The application can be initiated using the command \code{SeaClass()} or from RStudio's Addins dropdown menu. Note that categorical predictors with a large number of levels can be problematic for some of the application modules. Hence, categorical variables with only one level or more than 10 levels are often ignored. The user can manually adjust \code{max_num_of_levels} after downloading the R analysis code.
}
\examples{
\dontrun{
### Fake Data:
X <- matrix(rnorm(10000,0,1),ncol=10,nrow=1000)
X[1:100,1:2] <- X[1:100,1:2] + 3
Y <- c(rep(1,100), rep(0,900))
Fake_Data <- data.frame(Y = Y , X)
### SeaClass Rare Failure Data:
data("rareFailData")
SeaClass()}
}
